ai-research-copilot/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
numpy>=1.26.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
